backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML Core
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
sentence-transformers==2.2.2
openai==1.3.0
langchain==0.0.335
langchain-community==0.0.1

# NLP
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8
Pillow==10.1.0

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Document Processing
PyPDF2==3.0.1
pdfminer.six==20221105
python-docx==1.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Report Generation
jinja2==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
reportlab==4.0.7

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.25.2
